+++ script/tool/lib/src/common.dart
@@ -551,29 +551,27 @@ class ProcessRunner {
 }
 
 /// Finding version of [package] that is published on pub.
-///
-/// Note: you should manually close the [httpClient] when done using the finder.
 class PubVersionFinder {
-
   /// Constructor.
-  PubVersionFinder({@required this.package, this.pubHost = defaultPubHost, @required this.httpClient});
+  ///
+  /// Note: you should manually close the [httpClient] when done using the finder.
+  PubVersionFinder({this.pubHost = defaultPubHost, @required this.httpClient});
 
   /// The default pub host to use.
   static const String defaultPubHost = 'https://pub.dev';
 
-  /// The package name.
-  final String package;
-
   /// The pub host url, defaults to `https://pub.dev`.
   final String pubHost;
 
-  /// The http client, can override for testing.
+  /// The http client.
   ///
   /// You should manually close this client when done using this finder.
   final http.Client httpClient;
 
   /// Get the package version on pub.
-  Future<PubVersionFinderResponse> getPackageVersion() async {
+  Future<PubVersionFinderResponse> getPackageVersion(
+      {@required String package}) async {
+    assert(package != null && package.isNotEmpty);
     final Uri pubHostUri = Uri.parse(pubHost);
     final Uri url = pubHostUri.replace(path: '/packages/$package.json');
     final http.Response response = await httpClient.get(url);
@@ -591,14 +589,9 @@ class PubVersionFinder {
     }
     final List<Version> versions =
         (json.decode(response.body)['versions'] as List<dynamic>)
-            .map<Version>(
-                (final dynamic versionString) => Version.parse(versionString as String))
+            .map<Version>((final dynamic versionString) =>
+                Version.parse(versionString as String))
             .toList();
-
-    versions.sort((Version a, Version b) {
-      /// TODO(cyanglaz): Think about how to handle pre-release version with [Version.prioritize].
-      return b.compareTo(a);
-    });
 
     return PubVersionFinderResponse(
         versions: versions,
@@ -610,7 +603,12 @@ class PubVersionFinder {
 /// Represents a response for [PubVersionFinder].
 class PubVersionFinderResponse {
   /// Constructor.
-  const PubVersionFinderResponse({this.versions, this.result, this.httpResponse});
+  PubVersionFinderResponse({this.versions, this.result, this.httpResponse}) {
+    versions.sort((Version a, Version b) {
+      /// TODO(cyanglaz): Think about how to handle pre-release version with [Version.prioritize].
+      return b.compareTo(a);
+    });
+  }
 
   /// The versions found in [PubVersionFinder].
   ///
@@ -625,7 +623,7 @@ class PubVersionFinderResponse {
   final http.Response httpResponse;
 }
 
-/// An enum represents the result of [PubVersionFinder].
+/// An enum representing the result of [PubVersionFinder].
 enum PubVersionFinderResult {
   /// The version finder successfully found a version.
   success,
@@ -637,7 +635,7 @@ enum PubVersionFinderResult {
 
   /// The version finder failed to locate the package.
   ///
-  /// This is usually OK when the package is new.
+  /// This indicates the package is new.
   noPackageFound,
 }