+++ script/tool/test/common_test.dart
@@ -7,6 +7,7 @@ import 'dart:io';
 
 import 'package:args/command_runner.dart';
 import 'package:file/file.dart';
+import 'package:file/memory.dart';
 import 'package:flutter_plugin_tools/src/common.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/testing.dart';
@@ -20,11 +21,20 @@ import 'util.dart';
 void main() {
   RecordingProcessRunner processRunner;
   CommandRunner<void> runner;
+  FileSystem fileSystem;
+  Directory packagesDir;
+  Directory thirdPartyPackagesDir;
   List<String> plugins;
   List<List<String>> gitDirCommands;
   String gitDiffResponse;
 
   setUp(() {
+    fileSystem = MemoryFileSystem();
+    packagesDir = fileSystem.currentDirectory.childDirectory('packages');
+    thirdPartyPackagesDir = packagesDir.parent
+        .childDirectory('third_party')
+        .childDirectory('packages');
+
     gitDirCommands = <List<String>>[];
     gitDiffResponse = '';
     final MockGitDir gitDir = MockGitDir();
@@ -37,13 +47,13 @@ void main() {
       }
       return Future<ProcessResult>.value(mockProcessResult);
     });
-    initializeFakePackages();
+    initializeFakePackages(parentDir: packagesDir.parent);
     processRunner = RecordingProcessRunner();
     plugins = <String>[];
     final SamplePluginCommand samplePluginCommand = SamplePluginCommand(
       plugins,
-      mockPackagesDir,
-      mockFileSystem,
+      packagesDir,
+      fileSystem,
       processRunner: processRunner,
       gitDir: gitDir,
     );
@@ -52,35 +62,48 @@ void main() {
     runner.addCommand(samplePluginCommand);
   });
 
-  tearDown(() {
-    mockPackagesDir.deleteSync(recursive: true);
-  });
-
   test('all plugins from file system', () async {
-    final Directory plugin1 = createFakePlugin('plugin1');
-    final Directory plugin2 = createFakePlugin('plugin2');
+    final Directory plugin1 =
+        createFakePlugin('plugin1', packagesDirectory: packagesDir);
+    final Directory plugin2 =
+        createFakePlugin('plugin2', packagesDirectory: packagesDir);
     await runner.run(<String>['sample']);
     expect(plugins, unorderedEquals(<String>[plugin1.path, plugin2.path]));
   });
 
+  test('all plugins includes third_party/packages', () async {
+    final Directory plugin1 =
+        createFakePlugin('plugin1', packagesDirectory: packagesDir);
+    final Directory plugin2 =
+        createFakePlugin('plugin2', packagesDirectory: packagesDir);
+    final Directory plugin3 =
+        createFakePlugin('plugin3', packagesDirectory: thirdPartyPackagesDir);
+    await runner.run(<String>['sample']);
+    expect(plugins,
+        unorderedEquals(<String>[plugin1.path, plugin2.path, plugin3.path]));
+  });
+
   test('exclude plugins when plugins flag is specified', () async {
-    createFakePlugin('plugin1');
-    final Directory plugin2 = createFakePlugin('plugin2');
+    createFakePlugin('plugin1', packagesDirectory: packagesDir);
+    final Directory plugin2 =
+        createFakePlugin('plugin2', packagesDirectory: packagesDir);
     await runner.run(
         <String>['sample', '--plugins=plugin1,plugin2', '--exclude=plugin1']);
     expect(plugins, unorderedEquals(<String>[plugin2.path]));
   });
 
   test('exclude plugins when plugins flag isn\'t specified', () async {
-    createFakePlugin('plugin1');
-    createFakePlugin('plugin2');
+    createFakePlugin('plugin1', packagesDirectory: packagesDir);
+    createFakePlugin('plugin2', packagesDirectory: packagesDir);
     await runner.run(<String>['sample', '--exclude=plugin1,plugin2']);
     expect(plugins, unorderedEquals(<String>[]));
   });
 
   test('exclude federated plugins when plugins flag is specified', () async {
-    createFakePlugin('plugin1', parentDirectoryName: 'federated');
-    final Directory plugin2 = createFakePlugin('plugin2');
+    createFakePlugin('plugin1',
+        parentDirectoryName: 'federated', packagesDirectory: packagesDir);
+    final Directory plugin2 =
+        createFakePlugin('plugin2', packagesDirectory: packagesDir);
     await runner.run(<String>[
       'sample',
       '--plugins=federated/plugin1,plugin2',
@@ -91,8 +114,10 @@ void main() {
 
   test('exclude entire federated plugins when plugins flag is specified',
       () async {
-    createFakePlugin('plugin1', parentDirectoryName: 'federated');
-    final Directory plugin2 = createFakePlugin('plugin2');
+    createFakePlugin('plugin1',
+        parentDirectoryName: 'federated', packagesDirectory: packagesDir);
+    final Directory plugin2 =
+        createFakePlugin('plugin2', packagesDirectory: packagesDir);
     await runner.run(<String>[
       'sample',
       '--plugins=federated/plugin1,plugin2',
@@ -103,8 +128,10 @@ void main() {
 
   group('test run-on-changed-packages', () {
     test('all plugins should be tested if there are no changes.', () async {
-      final Directory plugin1 = createFakePlugin('plugin1');
-      final Directory plugin2 = createFakePlugin('plugin2');
+      final Directory plugin1 =
+          createFakePlugin('plugin1', packagesDirectory: packagesDir);
+      final Directory plugin2 =
+          createFakePlugin('plugin2', packagesDirectory: packagesDir);
       await runner.run(
           <String>['sample', '--base-sha=master', '--run-on-changed-packages']);
 
@@ -114,8 +141,10 @@ void main() {
     test('all plugins should be tested if there are no plugin related changes.',
         () async {
       gitDiffResponse = '.cirrus';
-      final Directory plugin1 = createFakePlugin('plugin1');
-      final Directory plugin2 = createFakePlugin('plugin2');
+      final Directory plugin1 =
+          createFakePlugin('plugin1', packagesDirectory: packagesDir);
+      final Directory plugin2 =
+          createFakePlugin('plugin2', packagesDirectory: packagesDir);
       await runner.run(
           <String>['sample', '--base-sha=master', '--run-on-changed-packages']);
 
@@ -124,8 +153,9 @@ void main() {
 
     test('Only changed plugin should be tested.', () async {
       gitDiffResponse = 'packages/plugin1/plugin1.dart';
-      final Directory plugin1 = createFakePlugin('plugin1');
-      createFakePlugin('plugin2');
+      final Directory plugin1 =
+          createFakePlugin('plugin1', packagesDirectory: packagesDir);
+      createFakePlugin('plugin2', packagesDirectory: packagesDir);
       await runner.run(
           <String>['sample', '--base-sha=master', '--run-on-changed-packages']);
 
@@ -137,8 +167,9 @@ void main() {
 packages/plugin1/plugin1.dart
 packages/plugin1/ios/plugin1.m
 ''';
-      final Directory plugin1 = createFakePlugin('plugin1');
-      createFakePlugin('plugin2');
+      final Directory plugin1 =
+          createFakePlugin('plugin1', packagesDirectory: packagesDir);
+      createFakePlugin('plugin2', packagesDirectory: packagesDir);
       await runner.run(
           <String>['sample', '--base-sha=master', '--run-on-changed-packages']);
 
@@ -151,9 +182,11 @@ packages/plugin1/ios/plugin1.m
 packages/plugin1/plugin1.dart
 packages/plugin2/ios/plugin2.m
 ''';
-      final Directory plugin1 = createFakePlugin('plugin1');
-      final Directory plugin2 = createFakePlugin('plugin2');
-      createFakePlugin('plugin3');
+      final Directory plugin1 =
+          createFakePlugin('plugin1', packagesDirectory: packagesDir);
+      final Directory plugin2 =
+          createFakePlugin('plugin2', packagesDirectory: packagesDir);
+      createFakePlugin('plugin3', packagesDirectory: packagesDir);
       await runner.run(
           <String>['sample', '--base-sha=master', '--run-on-changed-packages']);
 
@@ -168,10 +201,10 @@ packages/plugin1/plugin1/plugin1.dart
 packages/plugin1/plugin1_platform_interface/plugin1_platform_interface.dart
 packages/plugin1/plugin1_web/plugin1_web.dart
 ''';
-      final Directory plugin1 =
-          createFakePlugin('plugin1', parentDirectoryName: 'plugin1');
-      createFakePlugin('plugin2');
-      createFakePlugin('plugin3');
+      final Directory plugin1 = createFakePlugin('plugin1',
+          parentDirectoryName: 'plugin1', packagesDirectory: packagesDir);
+      createFakePlugin('plugin2', packagesDirectory: packagesDir);
+      createFakePlugin('plugin3', packagesDirectory: packagesDir);
       await runner.run(
           <String>['sample', '--base-sha=master', '--run-on-changed-packages']);
 
@@ -185,10 +218,11 @@ packages/plugin1/plugin1.dart
 packages/plugin2/ios/plugin2.m
 packages/plugin3/plugin3.dart
 ''';
-      final Directory plugin1 =
-          createFakePlugin('plugin1', parentDirectoryName: 'plugin1');
-      final Directory plugin2 = createFakePlugin('plugin2');
-      createFakePlugin('plugin3');
+      final Directory plugin1 = createFakePlugin('plugin1',
+          parentDirectoryName: 'plugin1', packagesDirectory: packagesDir);
+      final Directory plugin2 =
+          createFakePlugin('plugin2', packagesDirectory: packagesDir);
+      createFakePlugin('plugin3', packagesDirectory: packagesDir);
       await runner.run(<String>[
         'sample',
         '--plugins=plugin1,plugin2',
@@ -205,10 +239,10 @@ packages/plugin1/plugin1.dart
 packages/plugin2/ios/plugin2.m
 packages/plugin3/plugin3.dart
 ''';
-      final Directory plugin1 =
-          createFakePlugin('plugin1', parentDirectoryName: 'plugin1');
-      createFakePlugin('plugin2');
-      createFakePlugin('plugin3');
+      final Directory plugin1 = createFakePlugin('plugin1',
+          parentDirectoryName: 'plugin1', packagesDirectory: packagesDir);
+      createFakePlugin('plugin2', packagesDirectory: packagesDir);
+      createFakePlugin('plugin3', packagesDirectory: packagesDir);
       await runner.run(<String>[
         'sample',
         '--exclude=plugin2,plugin3',