+++ script/tool/CHANGELOG.md
@@ -3,6 +3,10 @@
 - Add `against-pub` flag for version-check, which allows the command to check version with pub.
 - Add `machine` flag for publish-check, which replaces outputs to something parsable by machines.
 
+## 0.1.1
+
+- Update the allowed third-party licenses for flutter/packages.
+
 ## 0.1.0+1
 
 - Re-add the bin/ directory.